+++ augur/tasks/git/util/facade_worker/facade_worker/facade00mainprogram.py
@@ -33,9 +33,9 @@ from .config import FacadeSession as FacadeSession
 from .utilitymethods import trim_commit, store_working_author, trim_author   
 from .analyzecommit import analyze_commit
 from .postanalysiscleanup import git_repo_cleanup
-from .facade05repofetch import git_repo_initialize, check_for_repo_updates, force_repo_updates, force_repo_analysis, git_repo_updates
+from .repofetch import git_repo_initialize, check_for_repo_updates, force_repo_updates, force_repo_analysis, git_repo_updates
 #.facade06analyze analysis moved to facade_tasks.py - IM 10/12/22
-from .facade07rebuildcache import nuke_affiliations, fill_empty_affiliations, invalidate_caches, rebuild_unknown_affiliation_and_web_caches
+from .rebuildcache import nuke_affiliations, fill_empty_affiliations, invalidate_caches, rebuild_unknown_affiliation_and_web_caches
 
 #from contributor_interfaceable.facade08contributorinterfaceable import ContributorInterfaceable
